+++ apps/api_admin/entities/code.py
@@ -1,7 +1,9 @@
-from typing import Optional
-from typing import List
+__all__ = [
+    'CreateCodeParser',
+    'CreateCodeSchema'
+]
 
-from fastapi import Body, Query
+from fastapi import Body
 from pydantic import BaseModel
 from pydantic.fields import Field