+++ apps/api_admin/entities/order.py
@@ -1,5 +1,7 @@
 __all__ = [
-    'ListOrderSchema', 'FilterOrderParser',
+    'ListOrderSchema',
+    'ReadOrderSchema',
+    'FilterOrderParser'
 ]
 
 from typing import Optional