+++ apps/api_admin/entities/question.py
@@ -1,5 +1,7 @@
 __all__ = [
-    'ReadQuestionSchema', 'ListQuestionSchema', 'FilterQuestionParser',
+    'ReadQuestionSchema',
+    'ListQuestionSchema',
+    'FilterQuestionParser'
 ]
 
 from typing import Optional