+++ apps/api_admin/entities/token.py
@@ -1,5 +1,6 @@
 __all__ = [
-    'CreateTokenParser', 'TokenSchema',
+    'CreateTokenParser',
+    'TokenSchema'
 ]
 
 from typing import Optional