+++ apps/api_admin/logics/__init__.py
@@ -3,3 +3,4 @@ from .order import OrderResolver
 from .question import QuestionResolver
 from .user import UserResolver
 from .code import create_sms_code, create_sms_code_task
+from .token import LoginResolver