+++ apps/api_admin/resources/code.py
@@ -1,13 +1,13 @@
 from fastapi import APIRouter, BackgroundTasks
 
-from extensions import Route, error_response, resp_success
+from extensions import error_schema, resp_success, Route
 from apps.api_admin.entities import CreateCodeSchema, CreateCodeParser
 from apps.api_admin.logics import create_sms_code, create_sms_code_task
 
-router = APIRouter(route_class=Route, responses=error_response)
+router = APIRouter(route_class=Route, responses=error_schema)
 
 
-@router.post('', response_model=CreateCodeSchema, status_code=201, responses=error_response)
+@router.post('', response_model=CreateCodeSchema, status_code=201)
 async def create_code(parser: CreateCodeParser, background_tasks: BackgroundTasks):
     """the api of create sms code