+++ apps/api_admin/resources/token.py
@@ -1,15 +1,15 @@
 from fastapi import APIRouter
 
-from extensions import Route, error_response, resp_success
-from apps.modules import TokenResolver
+from extensions import error_schema, resp_success, Route
 from apps.api_admin.entities import CreateTokenParser, TokenSchema
+from apps.api_admin.logics import LoginResolver
 
-router = APIRouter(route_class=Route, responses=error_response)
+router = APIRouter(route_class=Route, responses=error_schema)
 
 
 @router.post('', response_model=TokenSchema, status_code=201)
 async def create_token(parser: CreateTokenParser):
     """the api of login and get token"""
 
-    data = await TokenResolver.authentic(parser.cellphone, parser.code, 'AdminUser')
+    data = await LoginResolver.authentic(parser.cellphone, parser.code, 'AdminUser')
     return resp_success(data=data)