+++ apps/api_admin/resources/user.py
@@ -1,20 +1,18 @@
 from fastapi import APIRouter, Depends, Path
 
-from extensions import Route, error_response, resp_success, Pagination
-from apps.models import AdminUser
-from apps.modules import get_current_admin_user
+from extensions import error_schema, resp_success, Route, Pagination
+from apps.modules import current_admin_user
 from apps.api_admin.entities import (
     ReadUserSchema, ListUserSchema, UserSchema, PatchUserParser, FilterUserParser
 )
 from apps.api_admin.logics import UserResolver
 
-router = APIRouter(route_class=Route, responses=error_response)
+router = APIRouter(route_class=Route, responses=error_schema, dependencies=[current_admin_user])
 
 
 @router.get('/{u_id}', response_model=ReadUserSchema, status_code=200)
 async def read_user(
-        u_id: int = Path(..., description='用户id', ge=1),
-        admin_user: AdminUser = Depends(get_current_admin_user)
+        u_id: int = Path(..., description='用户id', ge=1)
 ):
     """the api of read one user"""
 
@@ -26,7 +24,6 @@ async def read_user(
 async def patch_user(
         u_id: int,
         parser: PatchUserParser,
-        admin_user: AdminUser = Depends(get_current_admin_user)
 ):
     """the api of update one user"""
 
@@ -36,8 +33,7 @@ async def patch_user(
 
 @router.get('', response_model=ListUserSchema, status_code=200)
 async def list_users(
-        parser: FilterUserParser = Depends(FilterUserParser),
-        admin_user: AdminUser = Depends(get_current_admin_user)
+        parser: FilterUserParser = Depends(FilterUserParser)
 ):
     """the api of read list users"""
 
@@ -45,7 +41,7 @@ async def list_users(
 
     query = UserResolver.list_users(payload)
     total = await query.count()
-    query = Pagination(query, payload['page'], payload.get('pagesize') or total).items()
+    query = Pagination(query, payload['page'], payload['pagesize']).items()
     result = await UserResolver.response_users(await query)
 
     return resp_success(data={'total': total, 'users': result})