+++ apps/modules/__init__.py
@@ -1,2 +1,2 @@
-from .token import TokenResolver, get_current_admin_user
+from .token import TokenResolver, get_current_admin_user, current_admin_user
 from .resource import ResourceOp