+++ apps/modules/token.py
@@ -3,18 +3,17 @@ from datetime import datetime, timedelta
 import traceback
 
 import jwt
-from fastapi import Request, Header
+from fastapi import Request, Header, Depends
 from tortoise.exceptions import OperationalError
 
 from config import Config
 from extensions import BadRequest, Unauthorized, NotFound, logger
-from redis_ext import SMSCodeRedis
 from apps.models import User, AdminUser
 
 
 class TokenResolver:
 
-    EXTEND_MODEL_MAP = {'AdminUser': AdminUser, 'User': User}
+    EXTEND_MODEL_MAP = {'AdminUser': AdminUser}
 
     @classmethod
     def encode_auth_token(cls, user_id: int, extend_user_id: int, extend_model: str):
@@ -42,45 +41,18 @@ class TokenResolver:
             raise BadRequest(message=str(e))
 
     @classmethod
-    async def authentic(cls, cellphone: str, code: str, extend_model: str):
-        """the entrance to get auth token"""
-        redis_client = SMSCodeRedis(cellphone)
-        if code == await redis_client.get():
-            user = await User.filter(cellphone=cellphone, is_delete=False).first()
-            if not user:
-                raise NotFound(f'User User.cellphone = {cellphone} is not exists or is deleted')
-
-            Model = cls.EXTEND_MODEL_MAP.get(extend_model)
-            if not Model:
-                raise BadRequest(message=f'Model {extend_model} error')
-
-            extend_user = await Model.filter(user_id=user.id, is_delete=False).first()
-
-            if not extend_user:
-                raise NotFound(message=f'{extend_model} User.cellphone = {cellphone} is not exists or is deleted')
-
-            token, login_time, token_expired = cls.encode_auth_token(user.id, extend_user.id, 'AdminUser')
-            extend_user.login_time = login_time
-            extend_user.token_expired = token_expired
-            await extend_user.save()
-
-            return {'token': token, 'user_id': user.id, 'extend_user_id': extend_user.id, 'extend_model': 'AdminUser'}
-        raise BadRequest(message='SMS Code error')
-
-    @classmethod
     async def query_user(cls, user_id: int, extend_user_id: str, extend_model: str, **kwargs):
         """select query user, extend_user"""
-
-        user = await User.filter(id=user_id, is_delete=False).first()
-        if not user:
-            raise NotFound(f'User User.id = {user_id} is not exists or is deleted')
 
         Model = cls.EXTEND_MODEL_MAP.get(extend_model)
         if not Model:
             raise BadRequest(message=f'Model {extend_model} error')
 
-        extend_user = await Model.filter(id=extend_user_id, user_id=user_id, is_delete=False).first()
+        user = await User.filter(id=user_id, is_delete=False).first()
+        if not user:
+            raise NotFound(f'User User.id = {user_id} is not exists or is deleted')
 
+        extend_user = await Model.filter(id=extend_user_id, user_id=user_id, is_delete=False).first()
         if not extend_user:
             raise NotFound(message=f'{extend_model} {extend_model}.id = {extend_user_id} is not exists or is deleted')
 
@@ -91,10 +63,14 @@ class TokenResolver:
         """check the token is expired or not"""
 
         a = not (extend_user.login_time and extend_user.token_expired)
-        b = data.get('token_expired') < now
-        c = data.get('login_time') < extend_user.login_time.timestamp()
-        d = data.get('token_expired') > extend_user.token_expired.timestamp()
-        if any([a, b, c, d]):
+        b = (now - data.get('token_expired')) > 1
+        # now > the expired_time in JWT (had expired)
+        c = extend_user.login_time.timestamp() - data.get('login_time') > 1
+        # the login time in db > the login time in JWT (there is a new login, this token is expired)
+        d = data.get('token_expired') - extend_user.token_expired.timestamp() > 1
+        # the expired time in db < the expired time in JWT (had expired)
+        _lis = [a, b, c, d]
+        if any(_lis):
             logger.error([a, b, c, d])
             raise Unauthorized(message='login expired, please login retry.')
 
@@ -113,8 +89,8 @@ class TokenResolver:
 
                 cls.check_timeout(extend_user, now, data)
 
-                request.state.extend_user = extend_user
                 request.state.user = user
+                request.state.extend_user = extend_user
                 return payload
         except jwt.PyJWTError:
             raise Unauthorized(message='login expired，please login retry！')
@@ -128,3 +104,6 @@ async def get_current_admin_user(request: Request, x_token: str = Header(..., de
 
     await TokenResolver.decode_token(request, x_token)
     return request.state.extend_user
+
+
+current_admin_user = Depends(get_current_admin_user)