+++ extensions/route.py
@@ -16,10 +16,10 @@ class Route(APIRoute):
 
             logger.info('start request'.center(60, '*'))
             logger.info(f'{request.method} {request.url}')
-            if (
-                    request.method in ['POST', 'PUT', 'PATCH']
-                    and request.headers.get('content-type') == 'application/json'
-            ):
+
+            methods = ['POST', 'PUT', 'PATCH']
+            content_type = request.headers.get('content-type')
+            if request.method in methods and content_type == 'application/json':
                 try:
                     params = await request.json()
                     if params: