+++ mixins/schema.py
@@ -14,4 +14,4 @@ class SchemaMixin(BaseModel):
 class FilterParserMixin(BaseModel):
 
     page: Optional[int] = Query(default=1, description='page', gte=1)
-    pagesize: Optional[int] = Query(default=None, description='pagesize', gte=1, lte=40)
+    pagesize: Optional[int] = Query(default=10, description='pagesize', gte=1, lte=40)